Purify icss utils
